Refactor Makefile to include utils.o in bench target; fix NH_COUNT_2D macro calculation and remove unused c2d_to_string function
